Moved some files out to shared directory
