+++ shared/agnostic/renderer/modal-dialog.js
